+++ src/main/java/com/daou/demo/organization/domain/Groups.java
@@ -3,7 +3,6 @@ package com.daou.demo.organization.domain;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 import javax.persistence.*;
 import java.util.List;
@@ -15,7 +14,7 @@ import java.util.List;
 public class Groups {
     @Id @Column(name = "group_id")
     @GeneratedValue
-    private Long id;
+    private Integer id;
     private String name;
     private String deptCode;
